backend/requirements.txt
fastapi
uvicorn
SQLAlchemy[asyncio]
aiosqlite
pydantic
python-dotenv
passlib[bcrypt]